Move clock to core-utils module
