+++ vector-app/src/fdroid/java/im/vector/app/fdroid/BackgroundSyncStarter.kt
@@ -18,10 +18,10 @@ package im.vector.app.fdroid
 
 import android.content.Context
 import im.vector.app.core.di.ActiveSessionHolder
-import im.vector.app.core.time.Clock
 import im.vector.app.fdroid.receiver.AlarmSyncBroadcastReceiver
 import im.vector.app.features.settings.BackgroundSyncMode
 import im.vector.app.features.settings.VectorPreferences
+import im.vector.lib.core.utils.timer.Clock
 import timber.log.Timber
 import javax.inject.Inject
