+++ vector/src/main/java/im/vector/app/features/crypto/keysbackup/setup/KeysBackupSetupSharedViewModel.kt
@@ -22,8 +22,8 @@ import androidx.lifecycle.ViewModel
 import com.nulabinc.zxcvbn.Strength
 import im.vector.app.R
 import im.vector.app.core.platform.WaitingViewData
-import im.vector.app.core.time.Clock
 import im.vector.app.core.utils.LiveEvent
+import im.vector.lib.core.utils.timer.Clock
 import org.matrix.android.sdk.api.MatrixCallback
 import org.matrix.android.sdk.api.listeners.ProgressListener
 import org.matrix.android.sdk.api.session.Session
