+++ vector/src/main/java/im/vector/app/features/home/ShouldShowUnverifiedSessionsAlertUseCase.kt
@@ -17,9 +17,9 @@
 package im.vector.app.features.home
 
 import im.vector.app.config.Config
-import im.vector.app.core.time.Clock
 import im.vector.app.features.VectorFeatures
 import im.vector.app.features.settings.VectorPreferences
+import im.vector.lib.core.utils.timer.Clock
 import javax.inject.Inject
 
 class ShouldShowUnverifiedSessionsAlertUseCase @Inject constructor(
