+++ vector/src/main/java/im/vector/app/features/home/room/detail/ChatEffectManager.kt
@@ -16,7 +16,7 @@
 
 package im.vector.app.features.home.room.detail
 
-import im.vector.app.core.time.Clock
+import im.vector.lib.core.utils.timer.Clock
 import org.matrix.android.sdk.api.session.events.model.toModel
 import org.matrix.android.sdk.api.session.room.model.message.MessageContent
 import org.matrix.android.sdk.api.session.room.model.message.MessageType
